refactor: lift repository variant state
